Updates validation to work with pre-Rails 5.1
`will_save_change_to_attribute?` looks like it may have been defined in
Rails 5.1.

If define, we will keep the `if` proc and only run the validation if the
attribute will be changed. Otherwise, it will always run the validation.

We may want to look into other dirty checks we could use if we want to
keep this support.