Refactor tests for SecureValidatable module
While working on
https://github.com/devise-security/devise-security/issues/62, I noticed
that there were a lot of edge cases that either weren't tested at all or
did not consider all cases. I wanted to address this issue before
finalizing my changes so that I have greater confidence that the changes
do not break existing functionality in the module.

This commit changes the following:
* Removes the unnecessary devise modules from the User class so that we
  can be sure only functionality related to this module is being tested.
* Removes a bunch of save commands, as writing to the DB slows tests
  down and most of the functionality could be tested without that.
* Cleans up language of test descriptions.
* Adds consistency to emails, passwords, and formatting for asserts.

This commit adds the following tests:
* Allow email to be blank if email is not required
* Fail if updating email from non-blank to blank
* Fail if password is blank on creation
* Fail if password is updated to blank
* Fail if password_confirmation does not match password or is blank
* Allow password_confirmation to be skipped
* Fail if new password equals existing password
* Fail if module is included in a class that doesn't support the
  validates API

This used https://github.com/devise-security/devise-security/pull/63 as
a starting point.

Co-authored-by: Kevin Olbrich <kevin.olbrich@gmail.com>