Fix translations for digits/symbols keys on password complexity
The following options should add the same error message:
* `{ digit: 1 }` and `{ digits: 1 }`
* `{ digit: 2 }` and `{ digits: 2 }`
* `{ symbol: 1 }` and `{ symbols: 1 }`
* `{ symbol: 2 }` and `{ symbols: 2 }`

As there were no translations for the `digits` and `symbols` keys by
default, this would result in a missing translation error.

Instead of having to add a separate locale key for `digits` and
`symbols`, I opted to map those keys to the correspoding singular word.
This way, as long as the singular word is translated, the plural word
will be too with no extra effort.

It could be argued that `{ digit: 2 }` and `{ symbol: 2 }` should not
be allowed due to them being singular words with a plural count. I
chose not to go down this route for two reasons:
* It would add additional error handling code to maintain as well as
  throw an exception for something that otherwise would not (unless
  the user enabled a specific setting).
* Based on the amount of contributed translation files, there are many
  non-native English speakers using the gem. This is a minor enough
  issue that they shouldn't be punished for it.