Disable mongoid on ruby head (#348)
Closes #344

Looks like mongoid and ruby-head aren't cooperating. Made #347 to address this later.

Not only does this clean up the config file, but the tests doesn't even get queued up on Github. This is vastly better than the previous approach, which queued up the jobs and they just completed really quickly.